--- kpoint_eri/resource_estimates/df/__init__.py
@@ -1,9 +1,0 @@
-from .build_eris_df import (
-        kpoint_df_eris,
-        build_eris_kpt
-        )
-from .double_factorize import (
-        double_factorize,
-        double_factorize_batched
-        )
-from .compute_lambda_df import compute_lambda
